requirements.txt
flask
numpy
tensorflow
pillow
waitress
urllib3